Changed type to unsigned short data

+++ lib/utils.cuh
@@ -44,4 +44,4 @@ __device__ void get_max(unsigned short *data, const unsigned long long N, unsign
 /*
     Function useful to compute the base to the power of exp
 */
-__device__ void power(unsigned base, unsigned exp, unsigned long *result);
+__device__ void power(unsigned base, unsigned exp, unsigned *result);
